Send/Flashcard reading in place of kanji if null
Before, if the kanji expression of the entry was null, as it's the
case for a katakana word, it would send (or add to flashcard) the string
"null".
The word "null" is useless to the user and might cause confusion, so
insted of sending the word null, we send the reading as the kanji
expression.

Before:
null [コラム]: <<meaning>>

Now:
コラム [コラム]: <<meaning>>